reorganize operators
