+++ brainpy/math/operators/__init__.py
@@ -3,21 +3,16 @@
 
 from . import multiplication
 from . import op_register
-from . import pre2syn as pre2syn_module
-from . import pre2post as pre2post_module
-from . import syn2post as syn2post_module
+from . import pre_syn_post as pre_syn_post_module
 from . import wrap_jax
 from . import spikegrad
 
 __all__ = multiplication.__all__ + op_register.__all__
-__all__ += pre2syn_module.__all__ + pre2post_module.__all__ + syn2post_module.__all__
-__all__ += wrap_jax.__all__ + spikegrad.__all__
+__all__ += pre_syn_post_module.__all__ + wrap_jax.__all__ + spikegrad.__all__
 
 
 from .multiplication import *
 from .op_register import *
-from .pre2syn import *
-from .pre2post import *
-from .syn2post import *
+from .pre_syn_post import *
 from .wrap_jax import *
 from .spikegrad import *
